Parallel merges implemented but a little kludgy
It was a simple matter of parallizing the merge loop in dlaed0. OpenMP
made it even easier. But I don't like how I had to allocate as many
workspaces as the GPU devices we use. It seems a bit wasteful (even with
64GB of RAM) and inconvenient.
At least, parallel merge works. Performance under 4 GPUs is about 2x
performance of 1 GPU. I want to do better, of course.

Aside: it's strange that MAGMA performs _worse_ when multiple GPUs are
thrown at it. Maybe the input size is a bit too small?

TODO
- Implement out-of-core merge. In particular, back-transformation should
  use both GPUs and CPUs. This way, a few merges at the top of
  divide-and-conquer trees can still benefit from the hybrid setting.
- Test LARGER matrices! 8192x8192 is still a kid's size.